Revert "Add DatePicker"
