+++ package.json
@@ -59,13 +59,11 @@
     "@wojtekmaj/enzyme-adapter-react-17": "^0.6.3",
     "classnames": "^2.3.1",
     "cssnano": "^5.0.7",
-    "date-fns": "^2.25.0",
     "enzyme": "^3.11.0",
     "enzyme-to-json": "^3.6.2",
     "lodash": "^4.17.21",
     "object-hash": "^2.2.0",
     "rc-slider": "^9.7.2",
-    "react-day-picker": "^7.4.10",
     "react-focus-lock": "^2.5.2",
     "react-merge-refs": "^1.1.0",
     "react-outside-click-handler": "^1.3.0",
@@ -73,8 +71,8 @@
     "serve": "^12.0.0",
     "storybook": "^6.3.8",
     "storybook-dark-mode": "^1.0.8",
-    "tippy.js": "^6.3.1",
     "uuid": "^8.3.2",
+    "tippy.js": "^6.3.1",
     "webpack-bundle-analyzer": "^4.4.2"
   },
   "devDependencies": {
@@ -124,10 +122,10 @@
     "react-dom": "^17.0.2",
     "rimraf": "^3.0.2",
     "style-loader": "^3.2.1",
-    "stylelint": "^14.0.1",
-    "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-property-sort-order-smacss": "^8.0.0",
-    "stylelint-config-standard": "^23.0.0",
+    "stylelint": "^13.13.1",
+    "stylelint-config-prettier": "^8.0.2",
+    "stylelint-config-property-sort-order-smacss": "^7.1.0",
+    "stylelint-config-standard": "^22.0.0",
     "stylelint-prettier": "^1.2.0",
     "ts-jest": "^27.0.4",
     "typescript": "^4.3.5",
