+++ src/components/Typography/Body.tsx
@@ -19,7 +19,7 @@ type BodyProps = {
   className?: string;
 } & React.HTMLAttributes<HTMLParagraphElement>;
 
-const Body = ({ children, className, type = "b1", ...rest }: BodyProps) => {
+const Body = ({ children, className, type, ...rest }: BodyProps) => {
   return (
     <p
       className={cx(
