--- src/styles/components/DatePicker.pcss
@@ -1,13 +1,0 @@
-@import "../prefix.pcss";
-@import "../theme.pcss";
-
-.$(prefix)--date-picker {
-  position: absolute;
-  z-index: 1000;
-  border: $border $focus;
-  background: black;
-  &__hidden {
-    visibility: hidden;
-    opacity: 0;
-  }
-}
