+++ src/styles/index.css
@@ -44,7 +44,6 @@
 @import "./components/Avatar.pcss";
 @import "./components/FloatingPanel.pcss";
 @import "./components/RangeInput.pcss";
-@import "./components/DatePicker.pcss";
 
 /* = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = =
  = Global customization of Bricks and Braces opionionated CSS. =
