+++ src/types.d.ts
@@ -2,5 +2,3 @@ type ForwardedRef<T> =
   | ((instance: T | null) => void)
   | MutableRefObject<T | null>
   | null;
-
-type Nullish = undefined | null;
